[♻️ REFACTOR] ServiceImpl 구조 변경
